Fixed warning.

+++ tio/linux-sample/example.c
@@ -143,7 +143,7 @@ void handler_init(
 tio_bool_t tio_action_handler(tio_action_t* action, tio_action_err_t* err, void* userdata)
 {
     printf("tio_action_handler called\n");
-    printf("%.*s: %.*s\n", action->alias_length, action->alias, action->action_name_length, action->action_name);
+    printf("%.*s: %.*s\n", (int)action->alias_length, action->alias,(int)action->action_name_length, action->action_name);
     return KII_TRUE;
 }
